% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello, World!}
\usage{
hello()
}
\arguments{
\item{argument.name}{Describe an argument here.}
}
\value{
This is the return value.
}
\description{
This is an example function named 'hello' which prints 'Hello, World!'.
}
\details{
You can learn more about package authoring with RStudio at:

   http://r-pkgs.had.co.nz/

   Some useful keyboard shortcuts for package authoring:

   Install Package:           'Ctrl + Shift + B'

   Check Package:             'Ctrl + Shift + E'

   Test Package:              'Ctrl + Shift + T'
}
