% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacalc.R
\name{meta.calc}
\alias{meta.calc}
\title{Meta Calculator}
\usage{
meta.calc(input.data, plot.title)
}
\arguments{
\item{input.data}{The data frame containing the data to be meta-analyzed.}

\item{plot.title}{The name of the meta object to be created.  Will be used as the title of the forest plot.}
}
\value{
A meta object.
}
\description{
Calculates meta-analysis statistics on multiple studies from data that has means and standard deviations of two groups for each study.
}
